create a new api for admin required and the user permission

+++ flask_blog_andy/flaskblog/posts/routes.py
@@ -5,22 +5,24 @@ from flaskblog.posts.forms import PostForm
 from flaskblog import db
 from flask_login import current_user, login_required
 
-
 posts = Blueprint('posts', __name__)
 
 
 @posts.route("/post/new", methods=['GET', 'POST'])
 @login_required
 def new_post():
-    form = PostForm()
-    if form.validate():
-        post = Post(title=form.title.data, content=form.content.data, author=current_user)
-        db.session.add(post)
-        db.session.commit()
-        flash('Your post has been created!', 'success')
+    if (current_user.can_add_post or current_user.is_admin):
+        form = PostForm()
+        if form.validate():
+            post = Post(title=form.title.data, content=form.content.data, author=current_user)
+            db.session.add(post)
+            db.session.commit()
+            flash('Your post has been created!', 'success')
+            return redirect(url_for('main.home'))
+        return render_template('create_post.html', title='New Post', form=form, legend='New Post')
+    else:
+        flash('You do not have permission to add posts.', 'danger')
         return redirect(url_for('main.home'))
-    return render_template('create_post.html', title='New Post',
-                           form=form, legend='New Post')
 
 
 @posts.route("/post/<int:post_id>")
@@ -32,32 +34,40 @@ def post(post_id):
 @posts.route("/post/<int:post_id>/update", methods=['GET', 'POST'])
 @login_required
 def update_post(post_id):
-    post = Post.query.get_or_404(post_id)
-    if post.author != current_user:
-        abort(403)
-    form = PostForm()
-    if form.validate():
-        post.title = form.title.data
-        post.content = form.content.data
-        db.session.commit()
-        flash('Your post has been updated!', 'success')
-        return redirect(url_for('posts.post', post_id=post.id))
-    elif request.method == 'GET':
-        form.title.data = post.title
-        form.content.data = post.content
-    return render_template('create_post.html', title='Update Post',
-                           form=form, legend='Update Post')
+    if (current_user.can_update_post or current_user.is_admin):
+        post = Post.query.get_or_404(post_id)
+        if post.author != current_user:
+            abort(403)
+        form = PostForm()
+        if form.validate():
+            post.title = form.title.data
+            post.content = form.content.data
+            db.session.commit()
+            flash('Your post has been updated!', 'success')
+            return redirect(url_for('posts.post', post_id=post.id))
+        elif request.method == 'GET':
+            form.title.data = post.title
+            form.content.data = post.content
+        return render_template('create_post.html', title='Update Post',
+                            form=form, legend='Update Post')
+    
+    else:
+        flash('You do not have permission to update posts.', 'danger')
+        return redirect(url_for('main.home'))
 
 
 @posts.route("/post/<int:post_id>/delete", methods=['POST'])
-@login_required
+@login_required 
 def delete_post(post_id):
-    post = Post.query.get_or_404(post_id)
-    if post.author != current_user:
-        abort(403)
-    db.session.delete(post)
-    db.session.commit()
-    flash('Your post has been deleted!', 'success')
-    return redirect(url_for('main.home'))
-
-
+    if (current_user.can_delete_post or current_user.is_admin):
+        post = Post.query.get_or_404(post_id)
+        if post.author != current_user:
+            abort(403)
+        db.session.delete(post)
+        db.session.commit()
+        flash('Your post has been deleted!', 'success')
+        return redirect(url_for('main.home'))
+    
+    else:
+        flash('You do not have permission to delete posts.', 'danger')
+        return redirect(url_for('main.home'))
